Transported scraper files to lib
